+++ package.json
@@ -31,7 +31,7 @@
     "@formatjs/intl": "^2.9.0",
     "@lukso/lsp-factory.js": "3.1.1",
     "@lukso/lsp-smart-contracts": "0.11.1",
-    "@lukso/web-components": "1.42.0",
+    "@lukso/web-components": "1.43.0",
     "@nuxt/devtools": "^0.7.4",
     "@nuxtjs/algolia": "^1.9.0",
     "@nuxtjs/device": "^3.1.0",
@@ -93,8 +93,5 @@
       "stylelint"
     ]
   },
-  "version": "1.0.0",
-  "resolutions": {
-    "@lukso/web-components": "portal:/Users/dominik/www/lukso/tools-web-components/package"
-  }
+  "version": "1.0.0"
 }
