+++ translations/en_US.json
@@ -61,7 +61,7 @@
   "token_details_own": "You own this token",
   "footer_need_help_url": "mailto:support@lukso.network",
   "web3_interface_error_message": "You tried to connect with a legacy profile that is no longer supported.",
-  "errors_invalid_address": "This is an invalid Universal Profile address",
+  "errors_invalid_address": "This is an invalid address",
   "header_connect": "Connect",
   "web3_connect_error": "There was an error while connecting to the browser extension.",
   "footer_privacy_policy_text": "Privacy Policy",
